Implemented Dyslexic font selection option
Created a settings page to include the ui to enable dyslexic font.
Used context manager to implement the font across all the pages and had to change the stylesheet implmentation to make it consistent across screens.
For future implementations, it is suggested to use the useFontContext() hook to get the selected font and pass it to the stylesheet using the getStyles() function. The implementation of the function can be referenced from the defaultCSS file of the Settings Screen.
Also added Async storage module in order to persist the font selection across app launches.
The custom dyslexic font does not work with the fontWeight css property (it is an issue with react-native) and hence bold font needs to be referenced by the selectedFontFamily.bold attribute. The font styles are defined in the index file of the Settings Screen.
Current issue with the dyslexic font is that the font used is larger than the default (Roboto) sans-serif font for the same value of font-size. Need to implement appropriate scaling.